能量机关v1.0，有基础的识别和预测，有大量优化空间
